split views.py into logical subfiles
